Persist BrowserID RSA keypair in a file, so server restarts don't affect it.
The built-in key generation is still completely insecure - there is a warning
about this now. Well, you can generate a proper key with OpenSSL and
provide it, but since everything is still very fragile, don't expect security.

Also, this commit replaces SortedDict with Python's built-in OrderedDict,
so newer Django versions won't complain.