Profile support and updated urls.py for newer Django compatibility.
The profile support is very basic. It's just that for some reason, without
OAuth2 and profile endpoints responding correctly, my Firefox seem
to have failed to work with sync. I don't know what was the issue,
just decided to implement it.

This OAuth2 implementation has no token revocation. Do NOT use it
to secure anything private. Profiles are not much, given that they
contain just the username and email. See the comments in the source
code for details.