Updates to make it work with Firefox Desktop 56/57 and Android 55.
I haven't used sync for some time, but decided I'd go and fix things.
So, I've upgraded all the dependencies (and properly pinned them all,
this time), including Django 1.11. I've updated the code and ensured
it works for my desktop and Android Firefoxes.

Note, to log in you will need to whitelist the FireSync base URL
in the "webchannel.allowObject.urlWhitelist" about:config setting.
Without this, you'll only get "No Such Channel" errors in the
browser console.

Note, this patch is somewhat dirty, as it uses base64-encoded
sync (my custom "x-sync-token") tokens only for Android, based
on the User-Agent header. This is probably needs to be done for
all x-sync-token usage, irrespective of the platform.