Added PyBrowserID cryptography.io-based module.
This file (cio.py) is not yet used by the project, and meant to be
placed as browserid/crypto/cio.py (with a patch to __init__.py).

Well, that's how I did it when I've upgraded to Python 3.x.

This is very dirty way of doing things, of course. Sorry about this.
I plan to fork or monkey patch PyBrowserID later.

Note: the code in this module is NOT audited. I'm not a cryptographer.
There is no crypto there, I'm just using cryptography.io functionality
to handle RSA keys (so, hopefully, it's all solid), but nonetheless
I could've made some mistake.